+++ packages/components/organisms/f-checkout/src/components/Checkout.vue
@@ -808,7 +808,7 @@ export default {
                 line1: {
                     required
                 },
-                city: {
+                locality: {
                     required
                 },
                 postcode: {