+++ packages/components/organisms/f-checkout/src/components/_tests/Checkout.test.js
@@ -40,7 +40,7 @@ const $v = {
         }
     },
     address: {
-        city: {
+        locality: {
             $dirty: false,
             required: true
         },
@@ -424,7 +424,7 @@ describe('Checkout', () => {
                         ...defaultCheckoutState,
                         isLoggedIn: true,
                         serviceType: CHECKOUT_METHOD_DELIVERY,
-                        address: { line1: 'Fleet Place House', postcode: 'EC4M 7RF', city: 'London' }
+                        address: { line1: 'Fleet Place House', postcode: 'EC4M 7RF', locality: 'London' }
                     }),
                     i18n,
                     localVue,
@@ -2039,7 +2039,7 @@ describe('Checkout', () => {
                             validFields: [
                                 'customer.mobileNumber',
                                 'address.line1',
-                                'address.city',
+                                'address.locality',
                                 'address.postcode'
                             ],
                             invalidFields: []
@@ -2098,7 +2098,7 @@ describe('Checkout', () => {
                             validFields: [
                                 'customer.mobileNumber',
                                 'address.line1',
-                                'address.city',
+                                'address.locality',
                                 'address.postcode'
                             ],
                             invalidFields: []
@@ -2162,7 +2162,7 @@ describe('Checkout', () => {
                     validFields: [
                         'customer.mobileNumber',
                         'address.line1',
-                        'address.city',
+                        'address.locality',
                         'address.postcode'
                     ],
                     invalidFields: []