+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-v0.72.0
+v0.73.0
 ------------------------------
 *March 18, 2021*
 
@@ -20,6 +20,16 @@ v0.72.0
 - Some tests that are not required as they're now covered in these new changes.
 
 
+v0.72.0
+------------------------------
+*March 18, 2021*
+
+### Changed
+- `address.city` to `address.locality` in checkout state.
+- `UPDATE_STATE` mutation to use `locality`.
+- `getAddress` to use locality.
+
+
 v0.71.0
 ------------------------------
 *March 16, 2021*
@@ -28,7 +38,7 @@ v0.71.0
 - `Checkout.vue` method clean up. Refactor `submitCheckout`.
 
 ### Added
-- Methods `handleFulfillableContext`, `processOrderNotFulfillable`, 
+- Methods `handleFulfillableContext`, `processOrderNotFulfillable`,
 `processOrderIsFulfillable` & `logInfo`.